+++ src/pages/Home/index.js
@@ -2,12 +2,11 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 
 import Header from '../../components/Header';
-import Modal from '../../components/Modal';
+import ProfileModal from '../../components/ProfileModal';
 
 import avatar from '../../assets/avatar.svg';
 import trash from '../../assets/trash.svg';
 import edit from '../../assets/edit.svg';
-import close from '../../assets/close.svg';
 
 import './styles.css';
 
@@ -35,43 +34,7 @@ function Home() {
                 <img src={avatar} alt="Avatar" className="avatar" />
               </button>
               {isModalVisible ? (
-                <Modal onClose={() => setIsModalvisible(false)}>
-                  <div className="home-modal">
-                    <div className="avatar-content">
-                      <img src={avatar} alt="Avatar" />
-                    </div>
-                    <div className="side-content">
-                      <div className="middle">
-                        <div className="bio">
-                          <h1>Juliano Reis</h1>
-                          <h2>Front-end Developer</h2>
-                          <strong>Idade</strong>
-                          <span>28</span>
-                          <strong>Tempo de empresa</strong>
-                          <span>2</span>
-                          <strong>Projetos que participou</strong>
-                          <span>Ifood</span>
-                        </div>
-                        <div className="footer-modal">
-                          <button type="button">
-                            <img src={trash} alt="Trash" />
-                          </button>
-                          <Link to="/update">
-                            <img src={edit} alt="Edit" />
-                          </Link>
-                        </div>
-                      </div>
-                      <div className="close-modal">
-                        <button
-                          type="button"
-                          onClick={() => setIsModalvisible(false)}
-                        >
-                          <img src={close} alt="Close" />
-                        </button>
-                      </div>
-                    </div>
-                  </div>
-                </Modal>
+                <ProfileModal onClose={() => setIsModalvisible(false)} />
               ) : null}
               <strong>Juliano Reis</strong>
               <span>Front-end Developer</span>
@@ -94,43 +57,7 @@ function Home() {
                 <img src={avatar} alt="Avatar" className="avatar" />
               </button>
               {isModalVisible ? (
-                <Modal onClose={() => setIsModalvisible(false)}>
-                  <div className="home-modal">
-                    <div className="avatar-content">
-                      <img src={avatar} alt="Avatar" />
-                    </div>
-                    <div className="side-content">
-                      <div className="middle">
-                        <div className="bio">
-                          <h1>Juliano Reis</h1>
-                          <h2>Front-end Developer</h2>
-                          <strong>Idade</strong>
-                          <span>28</span>
-                          <strong>Tempo de empresa</strong>
-                          <span>2</span>
-                          <strong>Projetos que participou</strong>
-                          <span>Ifood</span>
-                        </div>
-                        <div className="footer-modal">
-                          <button type="button">
-                            <img src={trash} alt="Trash" />
-                          </button>
-                          <Link to="/update">
-                            <img src={edit} alt="Edit" />
-                          </Link>
-                        </div>
-                      </div>
-                      <div className="close-modal">
-                        <button
-                          type="button"
-                          onClick={() => setIsModalvisible(false)}
-                        >
-                          <img src={close} alt="Close" />
-                        </button>
-                      </div>
-                    </div>
-                  </div>
-                </Modal>
+                <ProfileModal onClose={() => setIsModalvisible(false)} />
               ) : null}
               <strong>Juliano Reis</strong>
               <span>Front-end Developer</span>
@@ -153,43 +80,7 @@ function Home() {
                 <img src={avatar} alt="Avatar" className="avatar" />
               </button>
               {isModalVisible ? (
-                <Modal onClose={() => setIsModalvisible(false)}>
-                  <div className="home-modal">
-                    <div className="avatar-content">
-                      <img src={avatar} alt="Avatar" />
-                    </div>
-                    <div className="side-content">
-                      <div className="middle">
-                        <div className="bio">
-                          <h1>Juliano Reis</h1>
-                          <h2>Front-end Developer</h2>
-                          <strong>Idade</strong>
-                          <span>28</span>
-                          <strong>Tempo de empresa</strong>
-                          <span>2</span>
-                          <strong>Projetos que participou</strong>
-                          <span>Ifood</span>
-                        </div>
-                        <div className="footer-modal">
-                          <button type="button">
-                            <img src={trash} alt="Trash" />
-                          </button>
-                          <Link to="/update">
-                            <img src={edit} alt="Edit" />
-                          </Link>
-                        </div>
-                      </div>
-                      <div className="close-modal">
-                        <button
-                          type="button"
-                          onClick={() => setIsModalvisible(false)}
-                        >
-                          <img src={close} alt="Close" />
-                        </button>
-                      </div>
-                    </div>
-                  </div>
-                </Modal>
+                <ProfileModal onClose={() => setIsModalvisible(false)} />
               ) : null}
               <strong>Juliano Reis</strong>
               <span>Front-end Developer</span>
@@ -212,43 +103,7 @@ function Home() {
                 <img src={avatar} alt="Avatar" className="avatar" />
               </button>
               {isModalVisible ? (
-                <Modal onClose={() => setIsModalvisible(false)}>
-                  <div className="home-modal">
-                    <div className="avatar-content">
-                      <img src={avatar} alt="Avatar" />
-                    </div>
-                    <div className="side-content">
-                      <div className="middle">
-                        <div className="bio">
-                          <h1>Juliano Reis</h1>
-                          <h2>Front-end Developer</h2>
-                          <strong>Idade</strong>
-                          <span>28</span>
-                          <strong>Tempo de empresa</strong>
-                          <span>2</span>
-                          <strong>Projetos que participou</strong>
-                          <span>Ifood</span>
-                        </div>
-                        <div className="footer-modal">
-                          <button type="button">
-                            <img src={trash} alt="Trash" />
-                          </button>
-                          <Link to="/update">
-                            <img src={edit} alt="Edit" />
-                          </Link>
-                        </div>
-                      </div>
-                      <div className="close-modal">
-                        <button
-                          type="button"
-                          onClick={() => setIsModalvisible(false)}
-                        >
-                          <img src={close} alt="Close" />
-                        </button>
-                      </div>
-                    </div>
-                  </div>
-                </Modal>
+                <ProfileModal onClose={() => setIsModalvisible(false)} />
               ) : null}
               <strong>Juliano Reis</strong>
               <span>Front-end Developer</span>
