+++ src/pages/Home/styles.css
@@ -68,108 +68,6 @@ div.home-page .home-container .middle-content .naver-profile li .modal-button im
   height: 100%;
 }
 
-.modal {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 10;
-  background: rgba(0, 0, 0, 0.5);
-}
-
-.modal .modal-container {
-  background: #fff;
-  width: 1000px;
-  height: 500px;
-  margin: auto 0;
-}
-
-.modal .modal-container .home-modal {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-
-.modal .modal-container .home-modal .avatar-content img {
-  width: 500px;
-  height: 500px;
-}
-
-.modal .modal-container .home-modal .side-content {
-  display: flex;
-  width: 100%;
-  height: 100%;
-  justify-content: space-between;
-  margin: 0 auto;
-}
-
-.modal .modal-container .home-modal .side-content .middle {
-  display: flex;
-  flex-direction: column;
-  margin-left: 30px;
-}
-
-.modal .modal-container .home-modal .side-content .middle .bio h1 {
-  font-size: 24px;
-  font-weight: 600;
-  color: #000000;
-}
-
-.modal .modal-container .home-modal .side-content .middle .bio h2 {
-  font-size: 16px;
-  font-weight: normal;
-  color: #212121;
-  margin-top: 10px;
-}
-
-.modal .modal-container .home-modal .side-content .middle .bio strong {
-  font-size: 16px;
-  font-weight: 600;
-  color: #212121;
-  margin-top: 24px;
-}
-
-.modal .modal-container .home-modal .side-content .middle .bio  span {
-  font-size: 16px;
-  font-weight: normal;
-  color: #212121;
-  margin-top: 10px;
-}
-
-.modal .modal-container .home-modal .side-content .middle .footer-modal {
-  margin-top: 150px;
-}
-
-.modal .modal-container .home-modal .side-content .middle .footer-modal button {
-  margin-right: 16px;
-  border: 0;
-  background: unset;
-  cursor: pointer;
-}
-
-.modal .modal-container .home-modal .side-content .close-modal button {
-  transition: 0.2s;
-  margin-right: 21px;
-  border: 0;
-  background: none;
-  cursor: pointer;
-}
-
-.modal .modal-container .home-modal .side-content .close-modal button:hover {
-  opacity: 0.7;
-}
-
-.modal .modal-container .home-modal .side-content .bio {
-  display: flex;
-  flex-direction: column;
-}
-
 div.home-page .home-container .middle-content .naver-profile li strong {
   font-size: 16px;
   font-weight: 600;
@@ -244,70 +142,6 @@ div.home-page .home-container .middle-content .naver-profile .icons a img {
   div.home-page .home-container .middle-content .naver-profile li .modal-button img {
     width: 200px;
     height: 300px;
-  }
-
-  .modal {
-    height: 2000px;
-  }
-
-  .modal .modal-container {
-    width: 420px;
-    height: 251px;
-    display: flex;
-    align-items: center;
-  }
-
-  .modal .modal-container .home-modal {
-    display: flex;
-    align-items: center;
-  }
-
-  .modal .modal-container .home-modal .avatar-content img {
-    width: 227px;
-    height: 255px;
-  }
-
-  .modal .modal-container .home-modal .side-content .middle {
-    margin-left: 10px;
-  }
-
-  .modal .modal-container .home-modal .side-content .middle .bio h1 {
-    font-size: 12px;
-  }
-
-  .modal .modal-container .home-modal .side-content .middle .bio h2 {
-    font-size: 8px;
-    margin-top: 3px;
-  }
-
-  .modal .modal-container .home-modal .side-content .middle .bio strong {
-    font-size: 8px;
-    margin-top: 5px;
-  }
-
-  .modal .modal-container .home-modal .side-content .middle .bio  span {
-    font-size: 8px;
-    margin-top: 3px;
-  }
-
-  .modal .modal-container .home-modal .side-content .middle .footer-modal {
-    margin-top: 30px;
-  }
-
-  .modal .modal-container .home-modal .side-content .middle .footer-modal button img {
-    width: 10px;
-    height: 10px;
-  }
-
-  .modal .modal-container .home-modal .side-content .middle .footer-modal a img {
-    width: 10px;
-    height: 10px;
-  }
-
-  .modal .modal-container .home-modal .side-content .close-modal button img {
-    width: 10px;
-    height: 10px;
-    margin-left: 105px;
   }
 
   div.home-page .home-container .middle-content .naver-profile li strong {
